requirements.txt
Flask==3.1.2
psycopg2==2.9.11
python-dotenv==1.0.1

